handlers, draft
Former-commit-id: c6824b206980a8339941c8dd030f06c641117c8b
Former-commit-id: 9db131e15bc561137502d5e88f91d3047e289f8b

+++ preprocessor.py
@@ -1,5 +1,5 @@
 #! /usr/bin/python
-import os,sys,json
+import os,sys,json,time
 
 os.chdir(os.path.dirname(os.path.realpath(__file__)))
 
@@ -7,6 +7,10 @@ try:
     os.mkdir("gen")
 except OSError:
     pass
+
+if os.path.isfile("gen/prop.h") and os.path.getmtime("classes.json") < os.path.getmtime("gen/prop.h"):
+    print("Skipping preprocessor step. Configuration did not change.")
+    sys.exit(0)
 
 file=open("gen/prop.h","w")
 file.write("#pragma once\n")
